backend/requirements.txt
flask
flask-cors
google-generativeai
pdfplumber
gtts
requests
beautifulsoup4
python-dotenv
werkzeug
